move ancestry to separate component
also moving herdId to useEffect to prevent inifinite loop
